+++ crates/wdl-analysis/src/diagnostics.rs
@@ -12,6 +12,7 @@ use wdl_ast::TreeToken;
 use wdl_ast::Version;
 use wdl_ast::v1::PlaceholderOption;
 
+use crate::MeaninglessLintDirective;
 use crate::MisleadingDeclarationOrderRule;
 use crate::UnnecessaryFunctionCall;
 use crate::UnusedCallRule;
@@ -891,6 +892,15 @@ pub fn unnecessary_function_call(
         .with_rule(UnnecessaryFunctionCall::ID)
         .with_highlight(span)
         .with_label(label.to_string(), label_span)
+}
+
+/// Creates an "unnecessary function call" diagnostic.
+pub fn meaningless_lint_directive(rule: &str, span: Span) -> Diagnostic {
+    Diagnostic::note(format!(
+        "unnecessary `except` directive for lint rule `{rule}`"
+    ))
+    .with_rule(MeaninglessLintDirective::ID)
+    .with_highlight(span)
 }
 
 /// Generates a diagnostic error message when a placeholder option has a type