+++ crates/wdl-analysis/src/document/v1.rs
@@ -16,7 +16,9 @@ use petgraph::prelude::DiGraphMap;
 use url::Url;
 use wdl_ast::AstNode;
 use wdl_ast::AstToken;
+use wdl_ast::Comment;
 use wdl_ast::Diagnostic;
+use wdl_ast::Directive;
 use wdl_ast::Ident;
 use wdl_ast::Span;
 use wdl_ast::SupportedVersion;
@@ -42,6 +44,7 @@ use wdl_ast::v1::TaskDefinition;
 use wdl_ast::v1::TypeRef;
 use wdl_ast::v1::WorkflowDefinition;
 use wdl_ast::version::V1;
+use wdl_grammar::SyntaxKind;
 
 use super::Document;
 use super::DocumentData;
@@ -183,6 +186,18 @@ pub(crate) fn populate_document(
         "expected a supported V1 version"
     );
 
+    // Pre-populate all of the lint exceptions
+    document.analysis_diagnostics.add_exceptions(
+        ast.inner()
+            .descendants_with_tokens()
+            .flat_map(|d| {
+                d.into_token()
+                    .and_then(Comment::cast)
+                    .and_then(|c| c.directive().and_then(Directive::into_except))
+            })
+            .flatten(),
+    );
+
     // First start by processing imports, struct definitions, and enum definitions
     // This needs to be performed before processing tasks and workflows as
     // declarations might reference an imported or locally-defined struct or enum
@@ -242,7 +257,7 @@ fn add_namespace(
     let (uri, imported) = match resolve_import(graph, import, importer_index) {
         Ok(resolved) => resolved,
         Err(Some(diagnostic)) => {
-            document.analysis_diagnostics.push(diagnostic);
+            document.analysis_diagnostics.add(diagnostic);
             return;
         }
         Err(None) => return,
@@ -252,7 +267,7 @@ fn add_namespace(
     let ns = match import.namespace() {
         Some((ns, span)) => match document.namespaces.get(&ns) {
             Some(prev) => {
-                document.analysis_diagnostics.push(namespace_conflict(
+                document.analysis_diagnostics.add(namespace_conflict(
                     &ns,
                     span,
                     prev.span,
@@ -289,7 +304,7 @@ fn add_namespace(
             if !imported.data.structs.contains_key(from.text()) {
                 document
                     .analysis_diagnostics
-                    .push(struct_not_in_document(&from));
+                    .add(struct_not_in_document(&from));
                 return None;
             }
 
@@ -314,13 +329,13 @@ fn add_namespace(
                     if prev.namespace.is_none() {
                         document
                             .analysis_diagnostics
-                            .push(struct_conflicts_with_import(
+                            .add(struct_conflicts_with_import(
                                 aliased_name,
                                 prev.name_span,
                                 span,
                             ));
                     } else {
-                        document.analysis_diagnostics.push(imported_struct_conflict(
+                        document.analysis_diagnostics.add(imported_struct_conflict(
                             aliased_name,
                             span,
                             prev.name_span,
@@ -374,13 +389,13 @@ fn add_namespace(
                     if prev.namespace.is_none() {
                         document
                             .analysis_diagnostics
-                            .push(enum_conflicts_with_import(
+                            .add(enum_conflicts_with_import(
                                 aliased_name,
                                 prev.name_span,
                                 span,
                             ));
                     } else {
-                        document.analysis_diagnostics.push(imported_enum_conflict(
+                        document.analysis_diagnostics.add(imported_enum_conflict(
                             aliased_name,
                             span,
                             prev.name_span,
@@ -478,7 +493,7 @@ fn add_struct(document: &mut DocumentData, definition: &StructDefinition) {
         if !are_structs_equal(definition, &prev_def) {
             document
                 .analysis_diagnostics
-                .push(struct_conflicts_with_import(
+                .add(struct_conflicts_with_import(
                     name.text(),
                     name.span(),
                     prev.name_span,
@@ -486,7 +501,7 @@ fn add_struct(document: &mut DocumentData, definition: &StructDefinition) {
             return;
         }
     } else if let Some(ctx) = document.context(name.text()) {
-        document.analysis_diagnostics.push(name_conflict(
+        document.analysis_diagnostics.add(name_conflict(
             name.text(),
             Context::Struct(name.span()),
             ctx,
@@ -500,7 +515,7 @@ fn add_struct(document: &mut DocumentData, definition: &StructDefinition) {
         let name = decl.name();
         match members.get(name.text()) {
             Some(prev_span) => {
-                document.analysis_diagnostics.push(name_conflict(
+                document.analysis_diagnostics.add(name_conflict(
                     name.text(),
                     Context::StructMember(name.span()),
                     Context::StructMember(*prev_span),
@@ -534,7 +549,7 @@ fn add_enum(document: &mut DocumentData, definition: &EnumDefinition) {
     if version < SupportedVersion::V1(V1::Three) {
         document
             .analysis_diagnostics
-            .push(enum_not_supported(version, definition.name().span()));
+            .add(enum_not_supported(version, definition.name().span()));
         return;
     }
 
@@ -546,14 +561,14 @@ fn add_enum(document: &mut DocumentData, definition: &EnumDefinition) {
         if !are_enums_equal(definition, &prev_def) {
             document
                 .analysis_diagnostics
-                .push(enum_conflicts_with_import(
+                .add(enum_conflicts_with_import(
                     name.text(),
                     name.span(),
                     prev.name_span,
                 ))
         }
     } else if let Some(ctx) = document.context(name.text()) {
-        document.analysis_diagnostics.push(name_conflict(
+        document.analysis_diagnostics.add(name_conflict(
             name.text(),
             Context::Enum(name.span()),
             ctx,
@@ -567,7 +582,7 @@ fn add_enum(document: &mut DocumentData, definition: &EnumDefinition) {
         let name = variant.name();
         match variants.get(name.text()) {
             Some(prev_span) => {
-                document.analysis_diagnostics.push(name_conflict(
+                document.analysis_diagnostics.add(name_conflict(
                     name.text(),
                     Context::EnumVariant(name.span()),
                     Context::EnumVariant(*prev_span),
@@ -621,7 +636,7 @@ fn convert_ast_type(document: &mut DocumentData, ty: &wdl_ast::v1::Type) -> Type
     match converter.convert_type(ty) {
         Ok(ty) => ty,
         Err(diagnostic) => {
-            document.analysis_diagnostics.push(diagnostic);
+            document.analysis_diagnostics.add(diagnostic);
             Type::Union
         }
     }
@@ -716,7 +731,7 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
     // Check for a name conflict
     let name = definition.name();
     if let Some(ctx) = document.context(name.text()) {
-        document.analysis_diagnostics.push(name_conflict(
+        document.analysis_diagnostics.add(name_conflict(
             name.text(),
             Context::Task(name.span()),
             ctx,
@@ -791,14 +806,14 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
                     // a single implicit dependency edge; if so, it might be unused
                     let mut edges = graph.edges_directed(index, Direction::Outgoing);
 
-                    if let (Some(true), None) = (edges.next().map(|e| e.weight()), edges.next())
-                        && !decl.inner().is_rule_excepted(UnusedInputRule::ID)
-                    {
+                    if edges.all(|e| *e.weight()) {
                         let name = decl.name();
 
-                        document
-                            .analysis_diagnostics
-                            .push(unused_input(name.text(), name.span()).with_severity(severity));
+                        document.analysis_diagnostics.exceptable_add(
+                            unused_input(name.text(), name.span()).with_severity(severity),
+                            decl.inner(),
+                            &UnusedInputRule::EXCEPTABLE_NODES,
+                        );
                     }
                 }
             }
@@ -808,13 +823,12 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
                 if let Some(command_section_span) = command_section_span
                     && decl.inner().span().start() > command_section_span.end()
                     && let Some(severity) = config.diagnostics_config().misleading_declaration_order
-                    && !decl
-                        .inner()
-                        .is_rule_excepted(MisleadingDeclarationOrderRule::ID)
                 {
-                    document.analysis_diagnostics.push(
+                    document.analysis_diagnostics.exceptable_add(
                         misleading_declaration_order(name.text(), name.span())
                             .with_severity(severity),
+                        decl.inner(),
+                        &MisleadingDeclarationOrderRule::EXCEPTABLE_NODES,
                     );
                 }
 
@@ -829,19 +843,24 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
                 }
 
                 // Check for unused declaration
-                if let Some(severity) = config.diagnostics_config().unused_declaration {
-                    // Don't warn for environment variables as they are always implicitly used
-                    if decl.env().is_none()
-                        && graph
-                            .edges_directed(index, Direction::Outgoing)
-                            .next()
-                            .is_none()
-                        && !decl.inner().is_rule_excepted(UnusedDeclarationRule::ID)
-                    {
-                        document.analysis_diagnostics.push(
-                            unused_declaration(name.text(), name.span()).with_severity(severity),
-                        );
-                    }
+                let Some(severity) = config.diagnostics_config().unused_declaration else {
+                    continue;
+                };
+
+                let name = decl.name();
+
+                // Don't warn for environment variables as they are always implicitly used
+                if decl.env().is_none()
+                    && graph
+                        .edges_directed(index, Direction::Outgoing)
+                        .next()
+                        .is_none()
+                {
+                    document.analysis_diagnostics.exceptable_add(
+                        unused_declaration(name.text(), name.span()).with_severity(severity),
+                        decl.inner(),
+                        &UnusedDeclarationRule::EXCEPTABLE_NODES,
+                    );
                 }
             }
             TaskGraphNode::Output(decl) => {
@@ -1017,13 +1036,13 @@ fn add_workflow(document: &mut DocumentData, workflow: &WorkflowDefinition) -> b
     if let Some(prev) = &document.workflow {
         document
             .analysis_diagnostics
-            .push(duplicate_workflow(&name, prev.name_span));
+            .add(duplicate_workflow(&name, prev.name_span));
         return false;
     }
 
     // Check for a name conflict
     if let Some(ctx) = document.context(name.text()) {
-        document.analysis_diagnostics.push(name_conflict(
+        document.analysis_diagnostics.add(name_conflict(
             name.text(),
             Context::Workflow(name.span()),
             ctx,
@@ -1103,13 +1122,14 @@ fn populate_workflow(config: &Config, document: &mut DocumentData, workflow: &Wo
                         .edges_directed(index, Direction::Outgoing)
                         .next()
                         .is_none()
-                    && !decl.inner().is_rule_excepted(UnusedInputRule::ID)
                 {
                     let name = decl.name();
 
-                    document
-                        .analysis_diagnostics
-                        .push(unused_input(name.text(), name.span()).with_severity(severity));
+                    document.analysis_diagnostics.exceptable_add(
+                        unused_input(name.text(), name.span()).with_severity(severity),
+                        decl.inner(),
+                        &UnusedInputRule::EXCEPTABLE_NODES,
+                    );
                 }
             }
             WorkflowGraphNode::Decl(decl) => {
@@ -1135,10 +1155,11 @@ fn populate_workflow(config: &Config, document: &mut DocumentData, workflow: &Wo
                         .edges_directed(index, Direction::Outgoing)
                         .next()
                         .is_none()
-                        && !decl.inner().is_rule_excepted(UnusedDeclarationRule::ID)
                     {
-                        document.analysis_diagnostics.push(
+                        document.analysis_diagnostics.exceptable_add(
                             unused_declaration(name.text(), name.span()).with_severity(severity),
+                            decl.inner(),
+                            &UnusedDeclarationRule::EXCEPTABLE_NODES,
                         );
                     }
                 }
@@ -1258,7 +1279,7 @@ fn populate_workflow(config: &Config, document: &mut DocumentData, workflow: &Wo
                                     entry.insert(info);
                                 }
                                 IndexMapEntry::Occupied(entry) => {
-                                    document.analysis_diagnostics.push(name_conflict(
+                                    document.analysis_diagnostics.add(name_conflict(
                                         &name,
                                         Context::Name(NameContext::Decl(info.span)),
                                         Context::Name(NameContext::Decl(entry.get().span)),
@@ -1336,7 +1357,7 @@ fn add_conditional_statement(
                     let span = Span::new(else_span.start(), if_span.end() - else_span.start());
                     document
                         .analysis_diagnostics
-                        .push(else_if_not_supported(version, span));
+                        .add(else_if_not_supported(version, span));
                 }
                 ConditionalStatementClauseKind::Else => {
                     let span = clause
@@ -1345,7 +1366,7 @@ fn add_conditional_statement(
                         .span();
                     document
                         .analysis_diagnostics
-                        .push(else_not_supported(version, span));
+                        .add(else_not_supported(version, span));
                 }
                 ConditionalStatementClauseKind::If => {}
             }
@@ -1375,7 +1396,7 @@ fn add_conditional_statement(
         if !ty.is_coercible_to(&PrimitiveType::Boolean.into()) {
             document
                 .analysis_diagnostics
-                .push(if_conditional_mismatch(&ty, expr.span()));
+                .add(if_conditional_mismatch(&ty, expr.span()));
         }
     }
 }
@@ -1412,7 +1433,7 @@ fn add_scatter_statement(
         _ => {
             document
                 .analysis_diagnostics
-                .push(type_is_not_array(&ty, expr.span()));
+                .add(type_is_not_array(&ty, expr.span()));
             Type::Union
         }
     };
@@ -1457,7 +1478,7 @@ fn add_call_statement(
                     .get(input_name.text())
                     .map(|i| (i.ty.clone(), i.required))
                     .unwrap_or_else(|| {
-                        document.analysis_diagnostics.push(unknown_call_io(
+                        document.analysis_diagnostics.add(unknown_call_io(
                             &call_ty,
                             &input_name,
                             Io::Input,
@@ -1491,7 +1512,7 @@ fn add_call_statement(
                             if !matches!(expected_input_ty, Type::Union)
                                 && !name.ty.is_coercible_to(&expected_input_ty)
                             {
-                                document.analysis_diagnostics.push(call_input_type_mismatch(
+                                document.analysis_diagnostics.add(call_input_type_mismatch(
                                     &input_name,
                                     &expected_input_ty,
                                     &name.ty,
@@ -1501,7 +1522,7 @@ fn add_call_statement(
                         None => {
                             document
                                 .analysis_diagnostics
-                                .push(unknown_name(input_name.text(), input_name.span()));
+                                .add(unknown_name(input_name.text(), input_name.span()));
                         }
                     },
                 }
@@ -1511,7 +1532,7 @@ fn add_call_statement(
 
             for (name, input) in call_ty.inputs() {
                 if input.required && !seen.contains(name.as_str()) {
-                    document.analysis_diagnostics.push(missing_call_input(
+                    document.analysis_diagnostics.add(missing_call_input(
                         call_ty.kind(),
                         &target_name,
                         name,
@@ -1540,13 +1561,14 @@ fn add_call_statement(
         // Check for unused call
         if let Some(severity) = config.diagnostics_config().unused_call
             && !is_used
-            && !statement.inner().is_rule_excepted(UnusedCallRule::ID)
             && let Some(ty) = ty.as_call()
             && !ty.outputs().is_empty()
         {
-            document
-                .analysis_diagnostics
-                .push(unused_call(name.text(), name.span()).with_severity(severity));
+            document.analysis_diagnostics.exceptable_add(
+                unused_call(name.text(), name.span()).with_severity(severity),
+                statement.inner(),
+                &UnusedCallRule::EXCEPTABLE_NODES,
+            );
         }
 
         scope.insert(name.text(), name.span(), ty);
@@ -1574,7 +1596,7 @@ fn resolve_call_type(
         if namespace.is_some() {
             document
                 .analysis_diagnostics
-                .push(only_one_namespace(target.span()));
+                .add(only_one_namespace(target.span()));
             return None;
         }
 
@@ -1586,7 +1608,7 @@ fn resolve_call_type(
             None => {
                 document
                     .analysis_diagnostics
-                    .push(unknown_namespace(&target));
+                    .add(unknown_namespace(&target));
                 return None;
             }
         }
@@ -1599,7 +1621,7 @@ fn resolve_call_type(
     if namespace.is_none() && name.text() == workflow_name {
         document
             .analysis_diagnostics
-            .push(recursive_workflow_call(name.text(), name.span()));
+            .add(recursive_workflow_call(name.text(), name.span()));
         return None;
     }
 
@@ -1612,7 +1634,7 @@ fn resolve_call_type(
                 workflow.outputs.clone(),
             ),
             _ => {
-                document.analysis_diagnostics.push(unknown_task_or_workflow(
+                document.analysis_diagnostics.add(unknown_task_or_workflow(
                     namespace.map(|ns| ns.span),
                     name.text(),
                     name.span(),
@@ -1764,7 +1786,7 @@ fn populate_types(document: &mut DocumentData) {
                 return Ok(e.ty().cloned().unwrap_or(Type::Union));
             }
 
-            self.document.analysis_diagnostics.push(unknown_type(
+            self.document.analysis_diagnostics.add(unknown_type(
                 name,
                 Span::new(span.start() + self.offset, span.len()),
             ));
@@ -1826,7 +1848,7 @@ fn populate_types(document: &mut DocumentData) {
                     let def_name = definition.name();
                     let def_span = def_name.span();
                     let member_span = member.name().span();
-                    document.analysis_diagnostics.push(recursive_struct(
+                    document.analysis_diagnostics.add(recursive_struct(
                         def_name.text(),
                         Span::new(def_span.start() + s.offset, def_span.len()),
                         Span::new(member_span.start() + s.offset, member_span.len()),
@@ -1860,7 +1882,7 @@ fn populate_types(document: &mut DocumentData) {
                 if has_path_connecting(&graph, from_idx, to_idx, Some(&mut space)) {
                     let def_name = definition.name();
                     let def_span = def_name.span();
-                    document.analysis_diagnostics.push(recursive_enum(
+                    document.analysis_diagnostics.add(recursive_enum(
                         def_name.text(),
                         Span::new(def_span.start() + e.offset, def_span.len()),
                         match to_idx {
@@ -1898,7 +1920,7 @@ fn populate_types(document: &mut DocumentData) {
                             let span = label.span();
                             label.set_span(Span::new(span.start() + offset, span.len()));
                         }
-                        document.analysis_diagnostics.push(diagnostic);
+                        document.analysis_diagnostics.add(diagnostic);
                     }
                 }
             }
@@ -1918,7 +1940,7 @@ fn populate_types(document: &mut DocumentData) {
                                 let adjusted_span = Span::new(span.start() + e.offset, span.len());
                                 document
                                     .analysis_diagnostics
-                                    .push(non_literal_enum_value(adjusted_span));
+                                    .add(non_literal_enum_value(adjusted_span));
                                 Type::Union
                             }
                         }
@@ -1952,7 +1974,7 @@ fn populate_types(document: &mut DocumentData) {
                         document.enums[index].ty = Some(enum_ty.into());
                     }
                     Err(diagnostic) => {
-                        document.analysis_diagnostics.push(diagnostic);
+                        document.analysis_diagnostics.add(diagnostic);
                     }
                 }
             }
@@ -2175,7 +2197,18 @@ impl crate::types::v1::EvaluationContext for EvaluationContext<'_> {
     }
 
     fn add_diagnostic(&mut self, diagnostic: Diagnostic) {
-        self.document.analysis_diagnostics.push(diagnostic);
+        self.document.analysis_diagnostics.add(diagnostic);
+    }
+
+    fn exceptable_add_diagnostic<N: TreeNode + Exceptable>(
+        &mut self,
+        diagnostic: Diagnostic,
+        element: &N,
+        exceptable_nodes: &Option<&'static [SyntaxKind]>,
+    ) {
+        self.document
+            .analysis_diagnostics
+            .exceptable_add(diagnostic, element, exceptable_nodes);
     }
 }
 
@@ -2193,7 +2226,7 @@ fn type_check_expr(
     let actual = evaluator.evaluate_expr(expr).unwrap_or(Type::Union);
 
     if !matches!(expected, Type::Union) && !actual.is_coercible_to(expected) {
-        document.analysis_diagnostics.push(type_mismatch(
+        document.analysis_diagnostics.add(type_mismatch(
             expected,
             expected_span,
             &actual,
@@ -2208,7 +2241,7 @@ fn type_check_expr(
     {
         document
             .analysis_diagnostics
-            .push(non_empty_array_assignment(expected_span, expr.span()));
+            .add(non_empty_array_assignment(expected_span, expr.span()));
     }
 }