+++ crates/wdl-engine/src/eval/v1/task.rs
@@ -29,6 +29,7 @@ use tracing::enabled;
 use tracing::info;
 use tracing::warn;
 use walkdir::WalkDir;
+use wdl_analysis::Diagnostics;
 use wdl_analysis::Document;
 use wdl_analysis::diagnostics::Io;
 use wdl_analysis::diagnostics::multiple_type_mismatch;
@@ -645,7 +646,7 @@ impl Evaluator {
         let version = document.version().expect("document should have version");
 
         // Build an evaluation graph for the task
-        let mut diagnostics = Vec::new();
+        let mut diagnostics = Diagnostics::new();
         let graph =
             TaskGraphBuilder::default().build(version, &definition, &mut diagnostics, |name| {
                 document.struct_by_name(name).is_some() || document.enum_by_name(name).is_some()